Tests: Split plugin tests per production class and increase coverage
Move the tests out of the single JarhcGradlePluginTest, which covered
three production classes, into one test class per class under test:
JarhcGradlePluginTest, JarhcReportTaskTest, and JarhcRunnerTest. Add
tests for the previously uncovered branches: the worker parameter wiring
and deprecated-option warnings in JarhcReportTask, the missing-data-path,
empty-sections, non-positive-release, debug-logging and directory-error
paths in JarhcRunner, and the runtime-classpath convention in
JarhcGradlePlugin. Overall line coverage rises from 76% to about 91%.
